Moves tests to __tests__ folder
