+++ frontend/__tests__/components/Layout/Layout.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { shallow } from 'enzyme';
-import Layout from './Layout';
+import Layout from '../../../components/Layout/Layout';
 
 describe('round', () => {
   let mockProps;
