+++ frontend/__tests__/components/Nav/Nav.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { shallow } from 'enzyme';
-import Nav from './Nav';
+import Nav from '../../../components/Nav/Nav';
 
 describe('round', () => {
   let mockProps;
